HIST: unify HOST local with Central archive + per-hour gap repair (v13.12.0)
HOST local .bewehist was an independent 5Hz max-hold derivation with a fixed
-120..0 dB scale, so it could never be compared against Central's per-FFT-frame
archive. It now writes one row per FFT frame, quantized with the same truncating
formula and the same measured capture dB range, so the two files describe the
same rows.

Local copies are kept only until Central's copy is proven complete, then deleted.
The proof is local: a file is clean if the Central link never dropped and the
uplink queue never evicted a frame during its lifetime. Otherwise the file is
retained and, at the next hourly finalize, compared against Central via a new
HIST_STAT query (0x5C/0x5D) keyed on header start_utc rather than filename.
Only the missing tail rows are exported as an adjacent segment and uploaded;
the original is removed once coverage is proven, never on a single ACK.

Also fixes three real data-loss paths found while investigating a duplicate pair
on DGS-2 (2026-07-27):
  - Initial Central connect broadcast mission_sync before start_mux_adapter set
    central_sender_running_, so it was silently dropped. Central then refused to
    archive HIST for that station until something else re-synced the mission,
    while the HOST saw a clean link and deleted its only copy. Cost three hours
    of archive on DGS-2 and two on DGS-X.
  - archive_hist_on_live_start opened with "wb", so a reconnect landing in the
    same minute at the same centre frequency truncated the hour accumulated so
    far. Now collision-avoids with a suffix placed before "-LIVE", and finalize
    derives its name from the open file rather than stop.filename.
  - scan_orphans_enqueue swept MFS_HIST, which would truncate-overwrite Central's
    archive with a retained local copy on the next boot.

Viewer no longer truncates row_rate_hz to an integer (18.78 -> 18 inflated
Duration and Stop by 4.2% and disagreed with the hover tooltip).

+++ src/long_waterfall_view.cpp
@@ -801,8 +801,10 @@ void draw_info_modal(){
         if(s != std::string::npos) base = base.substr(s+1);
 
         int off_h = header_utc_offset(h);
-        uint32_t row_rate = (uint32_t)std::max(1.0f, h.row_rate_hz);
-        uint64_t dur_sec = hist_rows_from_header(h, sz) / row_rate;
+        // row_rate 를 정수로 자르면 안 된다 — Central 아카이브는 18.78Hz 같은 비정수라
+        // 18 로 잘리면 Duration/Stop 이 4.2% 부풀고, 같은 화면의 float 툴팁과 어긋난다.
+        float row_rate = std::max(1.0f, h.row_rate_hz);
+        uint64_t dur_sec = (uint64_t)(hist_rows_from_header(h, sz) / row_rate);
         uint64_t stop_utc = h.start_utc_unix + dur_sec;
 
         const float L = 110.f;
@@ -926,8 +928,8 @@ void draw_modal(FFTViewer& v, NetClient* cli){
         }
         const auto& h = g_open.hdr;
         int off_h = header_utc_offset(h);
-        uint32_t row_rate = (uint32_t)std::max(1.0f, h.row_rate_hz);
-        uint64_t dur_sec = (uint64_t)g_open.num_rows / row_rate;
+        float row_rate = std::max(1.0f, h.row_rate_hz);
+        uint64_t dur_sec = (uint64_t)((float)g_open.num_rows / row_rate);
         uint64_t stop_utc = h.start_utc_unix + dur_sec;
 
         unsigned fft_disp = h.fft_input_size > 0 ? h.fft_input_size : h.fft_size;
